# specs/001-distributed-banking-system/contracts/banking-api.openapi.yaml
openapi: 3.0.3
info:
  title: Distributed Banking API
  version: 0.1.0
  description: >
    Contract for authentication, balance, transfer, and admin observability endpoints
    for the distributed banking system course project.
servers:
  - url: https://api.example.local
    description: Local or cloud API base URL
tags:
  - name: Auth
  - name: Accounts
  - name: Transfers
  - name: Observability
paths:
  /auth/register:
    post:
      tags: [Auth]
      summary: Register a new user account
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '201':
          description: User registered
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '409':
          description: User already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /auth/login:
    post:
      tags: [Auth]
      summary: Login existing user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '401':
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /accounts/me/balance:
    get:
      tags: [Accounts]
      summary: Get current user account balance
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Balance retrieved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BalanceResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /transfers:
    post:
      tags: [Transfers]
      summary: Create transfer with idempotency
      security:
        - bearerAuth: []
      parameters:
        - in: header
          name: Idempotency-Key
          required: true
          schema:
            type: string
          description: Unique key to ensure retry-safe transfer execution
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TransferRequest'
      responses:
        '201':
          description: Transfer committed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TransferResponse'
        '200':
          description: Transfer replayed from idempotency record
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TransferResponse'
        '400':
          description: Invalid transfer request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Insufficient funds or invalid destination
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /admin/nodes/health:
    get:
      tags: [Observability]
      summary: List node health status by zone
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Node health list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NodeHealthListResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden (admin role required)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    ResponseMetadata:
      type: object
      required: [nodeId, zone, timestamp]
      properties:
        nodeId:
          type: string
        zone:
          type: string
        timestamp:
          type: string
          format: date-time
    RegisterRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 8
        role:
          type: string
          enum: [user, admin]
          default: user
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
    AuthResponse:
      type: object
      required: [accessToken, userId, role, metadata]
      properties:
        accessToken:
          type: string
        userId:
          type: string
        role:
          type: string
          enum: [user, admin]
        metadata:
          $ref: '#/components/schemas/ResponseMetadata'
    BalanceResponse:
      type: object
      required: [accountId, balance, currency, metadata]
      properties:
        accountId:
          type: string
        balance:
          type: number
          format: double
        currency:
          type: string
        metadata:
          $ref: '#/components/schemas/ResponseMetadata'
    TransferRequest:
      type: object
      required: [destinationAccountId, amount]
      properties:
        destinationAccountId:
          type: string
        amount:
          type: number
          format: double
          minimum: 0.01
        note:
          type: string
          maxLength: 256
    TransferResponse:
      type: object
      required: [transferId, status, sourceAccountId, destinationAccountId, amount, metadata]
      properties:
        transferId:
          type: string
        status:
          type: string
          enum: [COMMITTED, REJECTED]
        sourceAccountId:
          type: string
        destinationAccountId:
          type: string
        amount:
          type: number
          format: double
        message:
          type: string
        metadata:
          $ref: '#/components/schemas/ResponseMetadata'
    NodeHealth:
      type: object
      required: [nodeId, componentType, zone, status, lastObservedAt]
      properties:
        nodeId:
          type: string
        componentType:
          type: string
        zone:
          type: string
        status:
          type: string
          enum: [HEALTHY, UNHEALTHY, UNKNOWN]
        lastHealthAt:
          type: string
          format: date-time
          nullable: true
        lastObservedAt:
          type: string
          format: date-time
        message:
          type: string
    NodeHealthListResponse:
      type: object
      required: [nodes, metadata]
      properties:
        nodes:
          type: array
          items:
            $ref: '#/components/schemas/NodeHealth'
        metadata:
          $ref: '#/components/schemas/ResponseMetadata'
    ErrorResponse:
      type: object
      required: [code, message, metadata]
      properties:
        code:
          type: string
        message:
          type: string
        metadata:
          $ref: '#/components/schemas/ResponseMetadata'